requirements.txt
allure-pytest~=2.8.40
pytest~=6.2.2
pytest-xdist~=2.2.1
selenium~=3.141.0
webdriver-manager~=3.3.0
pytest-lazy-fixture~=0.6.3
requests~=2.25.1
Faker~=8.1.0
Appium-Python-Client~=1.1.0